requirements.txt
asgiref==3.8.1
astroid==3.3.5
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.35.90
botocore==1.35.90
bs4==0.0.2
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.9
Django==5.1.4
django-bower==5.2.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-countries==7.6.1
django-extensions==4.1
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-sslserver==0.22
django-storages==1.14.4
django-summernote==0.8.20.0
django-tenants==3.7.0
djangorestframework==3.15.2
findspark==2.0.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
idna==3.10
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
node==1.2.2
odict==1.9.0
pillow==11.0.0
platformdirs==4.3.6
plumber==1.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.4
setuptools==75.3.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
wheel==0.45.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.1.1
zope.lifecycleevent==5.0
zope.proxy==6.1

